requirements.txt
google-generativeai
google-search-results
requests